+++ src/types/Groups.ts
@@ -597,6 +597,10 @@ export enum GroupPermissionEnum {
     groupInstancePlusPortal = 'group-instance-plus-portal',
     groupInstancePlusPortalUnlocked = 'group-instance-plus-portal-unlocked',
     groupInstanceJoin = 'group-instance-join',
+    groupInstanceAgeGatedCreate = 'group-instance-age-gated-create',
+    groupInstanceAgeGatedJoin = 'group-instance-age-gated-join',
+    groupInstanceManage = 'group-instance-manage',
+    groupDefaultRoleManage = 'group-default-role-manage',
 }
 
 export type GroupPermissionsTags =